+++ packages/api/__tests__/Client.test.ts
@@ -11,7 +11,8 @@ const mock = new MockAdapter(axios) // set mock adapter on default axios instanc
 
 const ACCOUNT_ID = 'xZNQ1X'
 const CLIENT_ID = '218192'
-const testOptions: Options = { clientId: 'test-client-id' }
+const APPLICATION_CLIENT_ID = 'test-client-id'
+const testOptions: Options = {}
 
 const buildClientResponse = (clientResponseProperties: any = {}): any => ({
 	accounting_systemid: ACCOUNT_ID,
@@ -132,7 +133,7 @@ describe('@freshbooks/api', () => {
 	describe('Client', () => {
 		test('GET /accounting/account/<accountId>/users/clients?...searchQuery', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 			const response = `
 		  {
 		    "response": {
@@ -166,7 +167,7 @@ describe('@freshbooks/api', () => {
 		})
 		test('GET /accounting/account/<accountId>/users/clients', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 			const response = `
       {
         "response": {
@@ -200,7 +201,7 @@ describe('@freshbooks/api', () => {
 
 		test('GET /accounting/account/<accountId>/users/clients/<clientId>', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = buildMockClientJSONResponse()
 			mock.onGet(`/accounting/account/${ACCOUNT_ID}/users/clients/${CLIENT_ID}`).replyOnce(200, mockResponse)
@@ -214,7 +215,7 @@ describe('@freshbooks/api', () => {
 
 		test('POST /accounting/account/<accountId>/users/clients', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const clientModel = {
 				fName: 'Johnny',
@@ -249,7 +250,7 @@ describe('@freshbooks/api', () => {
 	})
 	test('PUT /accounting/account/<accountId>/users/clients/<clientId> (delete)', async () => {
 		const token = 'token'
-		const APIclient = new APIClient(token, testOptions)
+		const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 		const mockResponse = buildMockClientJSONResponse({ vis_state: 1 })
 
@@ -269,7 +270,7 @@ describe('@freshbooks/api', () => {
 
 	test('PUT /accounting/account/<accountId>/users/clients/<clientId>', async () => {
 		const token = 'token'
-		const APIclient = new APIClient(token, testOptions)
+		const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 		const mockResponse = buildMockClientJSONResponse()