+++ packages/api/__tests__/Expense.test.ts
@@ -11,6 +11,7 @@ const mock = new MockAdapter(axios) // set mock adapter on default axios instanc
 
 const ACCOUNT_ID = 'zDmNq'
 const EXPENSE_ID = '43221133'
+const APPLICATION_CLIENT_ID = 'test-client-id'
 
 const buildMockResponse = (expenseProperties: any = {}): string =>
 	JSON.stringify({
@@ -128,13 +129,13 @@ const buildMockRequest = (expenseProperties: any = {}): any => ({
 	},
 })
 
-const testOptions: Options = { clientId: 'test-client-id' }
+const testOptions: Options = {}
 
 describe('@freshbooks/api', () => {
 	describe('Expense', () => {
 		test('GET /accounting/account/<accountid>/expenses/expenses/<id>', async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = `{
 					"response":{
@@ -154,7 +155,7 @@ describe('@freshbooks/api', () => {
 
 		test('GET /accounting/account/<accountid>/expenses/expenses', async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = `{
 					"response":{
@@ -186,7 +187,7 @@ describe('@freshbooks/api', () => {
 
 		test('GET /accounting/account/<accountid>/expenses/expenses?include[]=category', async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const category = {
 				category: 'Accident Insurance',
@@ -235,7 +236,7 @@ describe('@freshbooks/api', () => {
 
 		test(`GET /accounting/account/<accountid>/expenses/expenses/search[expenseid]=${EXPENSE_ID}`, async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = `{
 					"response":{
@@ -271,7 +272,7 @@ describe('@freshbooks/api', () => {
 
 		test('POST /accounting/account/<accountid>/expenses/expenses', async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = `{
 					"response":{
@@ -293,7 +294,7 @@ describe('@freshbooks/api', () => {
 
 		test('PUT /accounting/account/<accountid>/expenses/expenses/<expenseid>', async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = `{
 					"response":{
@@ -317,7 +318,7 @@ describe('@freshbooks/api', () => {
 
 		test('PUT /accounting/account/<accountid>/expenses/expenses/<expenseid> (delete)', async () => {
 			const token = 'token'
-			const client = new APIClient(token, testOptions)
+			const client = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = `
             {"response":