+++ packages/api/__tests__/ServiceRate.test.ts
@@ -8,7 +8,8 @@ const mock = new MockAdapter(axios) // set mock adapter on default axios instanc
 
 const BUSINESS_ID = 12345
 const SERVICE_ID = 218192
-const testOptions: Options = { clientId: 'test-client-id' }
+const APPLICATION_CLIENT_ID = 'test-client-id'
+const testOptions: Options = {}
 
 const buildServiceRateResponse = (serviceRateResponseProperties: any = {}): any => ({
 	business_id: BUSINESS_ID,
@@ -39,7 +40,7 @@ describe('@freshbooks/api', () => {
 	describe('ServiceRate', () => {
 		test('GET /comments/business/<businessId>/service/<serviceId>/rate', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = buildMockResponse()
 
@@ -54,7 +55,7 @@ describe('@freshbooks/api', () => {
 
 		test('POST /comments/business/<businessId>/service/<serviceId>/rate', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockRequest = buildMockRequest()
 			const mockResponse = buildMockResponse()
@@ -69,7 +70,7 @@ describe('@freshbooks/api', () => {
 
 		test('PUT /comments/business/<businessId>/service/<serviceId>/rate', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockRequest = buildMockRequest()
 			const mockResponse = buildMockResponse()