+++ packages/api/__tests__/TimeEntry.test.ts
@@ -8,7 +8,8 @@ const mock = new MockAdapter(axios) // set mock adapter on default axios instanc
 
 const BUSINESS_ID = 12345
 const TIME_ENTRY_ID = 218192
-const testOptions: Options = { clientId: 'test-client-id' }
+const APPLICATION_CLIENT_ID = 'test-client-id'
+const testOptions: Options = {}
 
 const buildTimeEntryResponse = (timeEntryResponseProperties: any = {}): any => ({
 	id: TIME_ENTRY_ID,
@@ -67,7 +68,7 @@ describe('@freshbooks/api', () => {
 	describe('TimeEntry', () => {
 		test('GET /timetracking/business/<businessId>/time_entries', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 			const response = `{
 		    "meta": {
                 "total": 1,
@@ -97,7 +98,7 @@ describe('@freshbooks/api', () => {
 
 		test('GET /timetracking/business/<businessId>/time_entries/<timeEntryId>', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const mockResponse = buildMockTimeEntryJSONResponse()
 
@@ -112,7 +113,7 @@ describe('@freshbooks/api', () => {
 
 		test('POST /timetracking/business/<businessId>/time_entries', async () => {
 			const token = 'token'
-			const APIclient = new APIClient(token, testOptions)
+			const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 			const timeEntryModel = {
 				clientId: 9876,
@@ -151,7 +152,7 @@ describe('@freshbooks/api', () => {
 
 	test('PUT /timetracking/business/<businessId>/time_entries/<timeEntryId>', async () => {
 		const token = 'token'
-		const APIclient = new APIClient(token, testOptions)
+		const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 		const mockResponse = buildMockTimeEntryJSONResponse({
 			started_at: '2020-01-21T10:00:00.000Z',
@@ -186,7 +187,7 @@ describe('@freshbooks/api', () => {
 
 	test('DELETE /timetracking/business/<businessId>/time_entries/<timeEntryId>', async () => {
 		const token = 'token'
-		const APIclient = new APIClient(token, testOptions)
+		const APIclient = new APIClient(APPLICATION_CLIENT_ID, token, testOptions)
 
 		mock.onDelete(`/timetracking/business/${BUSINESS_ID}/time_entries/${TIME_ENTRY_ID}`).replyOnce(204, {})